pyproject.toml
[project]
name = "finetuning-on-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain-google-genai>=3.0.3",
    "python-dotenv>=1.2.1",
]
